# operations/operations-rest-services/src/main/java/com/sbcc/operations/service/api/CaseactivityApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (1.0.11).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.sbcc.operations.service.api;

import com.sbcc.operations.service.model.CaseActivityEventRequest;
import com.sbcc.operations.service.model.CaseActivityEventResponse;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2018-01-13T14:24:12.340Z")

@Api(value = "caseactivity", description = "the caseactivity API")
public interface CaseactivityApi {

    @ApiOperation(value = "submit Case Activity event", nickname = "caseActivityEvent", notes = "Submits a Case Activity event that will be processed by the SBCC ", response = CaseActivityEventResponse.class, tags={ "application", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "event will be processed", response = CaseActivityEventResponse.class),
        @ApiResponse(code = 400, message = "bad input parameter") })
    @RequestMapping(value = "/caseactivity",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.PUT)
    ResponseEntity<CaseActivityEventResponse> caseActivityEvent(@ApiParam(value = "case activity event that has to be processed by SBCC" ,required=true )  @Valid @RequestBody CaseActivityEventRequest caseActivityEventRequest);

}